+++ App_Data/jobs/continuous/cancelGMOAuthorization/run.ts
@@ -1,12 +1,13 @@
 /**
- * ムビチケ着券オーソリキャンセル
+ * GMO仮売上キャンセル
  *
  * @ignore
  */
+
 import * as sskts from '@motionpicture/sskts-domain';
 import * as mongoose from 'mongoose';
 
-import mongooseConnectionOptions from '../../mongooseConnectionOptions';
+import mongooseConnectionOptions from '../../../../mongooseConnectionOptions';
 
 (<any>mongoose).Promise = global.Promise;
 mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
@@ -15,7 +16,7 @@ let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
 const INTERVAL_MILLISECONDS = 1000;
-const queueAdapter = sskts.adapter.queue(mongoose.connection);
+const taskAdapter = sskts.adapter.task(mongoose.connection);
 
 setInterval(
     async () => {
@@ -26,7 +27,9 @@ setInterval(
         count += 1;
 
         try {
-            await sskts.service.queue.executeCancelMvtkAuthorization()(queueAdapter);
+            await sskts.service.task.executeByName(
+                sskts.factory.taskName.CancelGMOAuthorization
+            )(taskAdapter, mongoose.connection);
         } catch (error) {
             console.error(error.message);
         }