+++ bin/importFilms/index.ts
@@ -9,11 +9,11 @@ import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../mongooseConnectionOptions';
 
-const debug = createDebug('sskts-api:*');
+const debug = createDebug('sskts-jobs:*');
 
 async function main() {
     debug('connecting mongodb...');
-    mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions);
+    mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
     const theaterAdapter = sskts.adapter.theater(mongoose.connection);
     const filmAdapter = sskts.adapter.film(mongoose.connection);