+++ bin/importPerformances/index.js
@@ -18,7 +18,7 @@ const createDebug = require("debug");
 const moment = require("moment");
 const mongoose = require("mongoose");
 const mongooseConnectionOptions_1 = require("../../mongooseConnectionOptions");
-const debug = createDebug('sskts-api:*');
+const debug = createDebug('sskts-jobs:*');
 function main() {
     return __awaiter(this, void 0, void 0, function* () {
         debug('connecting mongodb...');