+++ bin/watchSettleMvtkAuthorizationQueue/index.ts
@@ -13,7 +13,7 @@ import mongooseConnectionOptions from '../../mongooseConnectionOptions';
 const debug = createDebug('sskts-jobs:bin:watchSettleMvtkAuthorizationQueue');
 
 (<any>mongoose).Promise = global.Promise;
-mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions);
+mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
 let count = 0;